tests/wpt/tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/a-user-click-during-pageshow.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
<script>
"use strict";
promise_test(async t => {
  const sentinelIframe = await setupSentinelIframe(t);
  const startingHistoryLength = history.length;

  const code = `
    window.onpageshow = () => {
      const a = document.createElement("a");
      a.href = "/common/blank.html?thereplacement";
      a.id = "the-anchor";
      a.textContent = "needs to have content to be clickable";
      document.body.append(a);
      parent.test_driver.click(a);
    };
  `;

  const startURL = "resources/slow-code-injector.html?pipe=sub(none)&code=" + encodeURIComponent(code);
  const afterReplacementURL = "/common/blank.html?thereplacement";
  const iframe = insertIframe(t, startURL);

  assert_equals(history.length, startingHistoryLength, "Inserting the under-test iframe must not change history.length");

  await waitForLoadAllowingIntermediateLoads(t, iframe, afterReplacementURL);
  assert_equals(history.length, startingHistoryLength + 1, "history.length must change after waiting for the load");
}, "User click on <a> during the pageshow event must NOT replace");
</script>
